+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/project/extensions/shared/features/tasks/GenerateProjectDataTask.kt
@@ -1,5 +1,6 @@
 package com.javiersc.hubdle.project.extensions.shared.features.tasks
 
+import com.android.build.gradle.internal.tasks.factory.dependsOn
 import com.javiersc.hubdle.project.extensions._internal.allKotlinSrcDirsWithoutBuild
 import com.javiersc.hubdle.project.extensions._internal.kotlinGeneratedSrcDirs
 import com.javiersc.hubdle.project.extensions._internal.kotlinSrcDirsWithoutBuild
@@ -9,6 +10,7 @@ import com.javiersc.kotlin.stdlib.remove
 import java.io.File
 import javax.inject.Inject
 import org.gradle.api.Project
+import org.gradle.api.Task
 import org.gradle.api.file.Directory
 import org.gradle.api.file.FileTree
 import org.gradle.api.file.RegularFileProperty
@@ -27,9 +29,12 @@ import org.gradle.api.tasks.PathSensitivity
 import org.gradle.api.tasks.SkipWhenEmpty
 import org.gradle.api.tasks.SourceTask
 import org.gradle.api.tasks.TaskAction
+import org.gradle.api.tasks.TaskCollection
 import org.gradle.api.tasks.TaskProvider
 import org.gradle.kotlin.dsl.property
 import org.gradle.kotlin.dsl.register
+import org.gradle.kotlin.dsl.withType
+import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
 
 @CacheableTask
 public open class GenerateProjectDataTask
@@ -143,7 +148,12 @@ constructor(
             val projectGroup = project.provider { "${project.group}" }
             val projectName = project.provider { project.name }
             val projectVersion = project.provider { "${project.version}" }
+            val assemble: TaskProvider<Task> = project.tasks.named(BasePlugin.ASSEMBLE_TASK_NAME)
+            val kotlinCompile: TaskCollection<KotlinCompile> =
+                project.tasks.withType<KotlinCompile>()
             val generateProjectDataTask = project.tasks.register<GenerateProjectDataTask>(NAME)
+            val prepareKotlinIdeaImport = project.tasks.maybeCreate("prepareKotlinIdeaImport")
+
             generateProjectDataTask.configure {
                 it.source(project.allKotlinSrcDirsWithoutBuild)
                 it.packageName.set(packageName)
@@ -157,6 +167,9 @@ constructor(
                     outputDir.get().file(fileName).asFile
                 }
             }
+            prepareKotlinIdeaImport.dependsOn(generateProjectDataTask)
+            kotlinCompile.configureEach { it.dependsOn(generateProjectDataTask) }
+            assemble.dependsOn(generateProjectDataTask)
             return generateProjectDataTask
         }